Moved the ResolveFile structure to a new package
